# Invoke-Formatter: Skip VariableAnalysis, which is not used by Formatter rules - this yields a 25%-50% performance improvement