## PR Summary

Formatter rules do not call APIs from VariableAnalysis, therefore skipping its initialization makes formatting much faster.
For PowerShell's 3000 line [build.psm1](https://github.com/PowerShell/PowerShell/blob/master/build.psm1) file, this changes execution time from 6 seconds to 5 seconds for a cold run and for a warm run from 2 seconds to 1.5 seconds. For a 10 line script, a cold run run goes down from 200ms to 100ms but due to it's small size, there was no measurable difference for warm runs (15ms).

## PR Checklist

- [x] [PR has a meaningful title](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
    - Use the present tense and imperative mood when describing your changes
- [x] [Summarized changes](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
- [x] [Change is not breaking](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#making-breaking-changes)
- [x] [Make sure all `.cs`, `.ps1` and `.psm1` files have the correct copyright header](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
- [ ] Make sure you've added a new test if existing tests do not effectively test the code changed and/or updated documentation
- [x] This PR is ready to merge and is not [Work in Progress](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---work-in-progress).
    - If the PR is work in progress, please add the prefix `WIP:` to the beginning of the title and remove the prefix when the PR is ready.